Create Desktop_Clean.py
The main changes from the previous code are as follows:

The file extensions for shortcuts and URLs are .webloc on macOS, so I added this to the list of file extensions to be removed.

The folder separator on macOS is /, so I used os.path.join() to create the new directory path with the correct separator.

Other than these changes, the code remains largely the same. Again, note that this code may not capture all possible files on your desktop and may need to be customized based on your specific needs.