Create Meta_Data_FolderTarget.py
The program uses the os and datetime modules to loop through all the files in the specified directory and add metadata for creation date using the setfile command line tool on macOS.

The PIL module is used to extract the EXIF data from the images. The TAGS dictionary maps EXIF tag numbers to their corresponding tag names, and the GPSTAGS dictionary maps GPS tag numbers to their corresponding tag names.

For each image, the program checks if it has an EXIF tag called DateTimeOriginal that contains the creation date. If it does, the program extracts the creation date, formats it as a string, and uses the setfile command line tool to add the creation date metadata to the file.

Note that the setfile command line tool is specific to macOS, and this program may need to be modified for use on other operating systems.