Create dvd_to_mp4.py
This program uses the pydvdread library to read the contents of the DVD and the subprocess module to run FFmpeg to perform the conversion. First, the DVD is mounted and the list of titles is obtained. The longest title (which is assumed to be the main feature) is then selected. The paths for the DVD and output file are obtained, and an FFmpeg command is constructed to perform the conversion. The subprocess.run function is used to run the FFmpeg command, and the resulting MP4 file is saved to the current working directory as "output.mp4". Finally, the DVD is ejected.

You will need to have the pydvdread and ffmpeg libraries installed to run this code. You can run this program in a Python environment or IDE such as PyCharm, Jupyter Notebook, or Spyder by saving the code as a Python file (e.g. dvd_to_mp4_converter.py) and running it. Note that you may need to adjust the path to the DVD drive (/dev/sr0 in this example) depending on your operating system and setup.