Create PrimeHarmonicsinAnalyticContinuation.py
This program utilizes TensorFlow to incorporate a neural network into the prime distribution code. It trains the neural network to learn from the output of the prime distribution function and make adjustments to its behavior accordingly. The mutate_distribution function is called after each epoch of training to introduce random mutations into the weights of the neural network, allowing it to explore different patterns in the prime distribution.

This code outputs the x values (the range of integers from 1 to 1000) and the y values (the predicted values of the prime distribution function for each x value).